+++ src/modules/equipments/views/edit.vue
@@ -44,14 +44,16 @@
       },
 
       onRemove() {
-        EquipmentsService.remove(this.equipment.id).then(() => {
-          this.setAlert({
-            message: 'Equipamento removido com sucesso!',
-            type: 'success',
-            from: this.$route.path,
-          });
+        EquipmentsService.confirmRemoval(this.equipment).then(() => {
+          EquipmentsService.remove(this.equipment.id).then(() => {
+            this.setAlert({
+              message: 'Equipamento removido com sucesso!',
+              type: 'success',
+              from: this.$route.path,
+            });
 
-          this.$router.go({ path: '/equipments' });
+            this.$router.go({ path: '/equipments' });
+          });
         });
       },
