+++ src/modules/permissions/views/permission-groups/edit.vue
@@ -44,14 +44,16 @@
       },
 
       onRemove() {
-        PermissionGroupsService.remove(this.permissionGroup.id).then(() => {
-          this.setAlert({
-            message: 'Grupo de permissão removido com sucesso!',
-            type: 'success',
-            from: this.$route.path,
-          });
+        PermissionGroupsService.confirmRemoval(this.permissionGroup).then(() => {
+          PermissionGroupsService.remove(this.permissionGroup.id).then(() => {
+            this.setAlert({
+              message: 'Grupo de permissão removido com sucesso!',
+              type: 'success',
+              from: this.$route.path,
+            });
 
-          this.$router.go({ path: '/permission_groups' });
+            this.$router.go({ path: '/permission_groups' });
+          });
         });
       },
