+++ src/modules/vehicles/views/edit.vue
@@ -44,14 +44,16 @@
       },
 
       onRemove() {
-        VehiclesService.remove(this.vehicle.id).then(() => {
-          this.setAlert({
-            message: 'Veículo removido com sucesso!',
-            type: 'success',
-            from: this.$route.path,
-          });
+        VehiclesService.confirmRemoval(this.vehicle).then(() => {
+          VehiclesService.remove(this.vehicle.id).then(() => {
+            this.setAlert({
+              message: 'Veículo removido com sucesso!',
+              type: 'success',
+              from: this.$route.path,
+            });
 
-          this.$router.go({ path: '/vehicles' });
+            this.$router.go({ path: '/vehicles' });
+          });
         });
       },
